Main Level Updated
Added a door between the exit door and the hallway

Added a new extinguisher prefab (now there are two prefabs one for the pickup and one for the thrown, when I added the thrown script to just the one I caused all the extinguishers on the ground to turn to red boxes. Keep the sphere collider I had problems with the capsule collider as it would collide with the player when thrown at certain angles).

Organized the assets, walls, doors, etc. into folders in the main’ level scene

-Kevin Williams